Add task solution - Code refactoring
