+++ src/App.jsx
@@ -1,28 +1,11 @@
 import './App.scss';
-import { PostList } from './components/PostList/PostList';
-
-import postsFromServer from './api/posts.json';
-import commentsFromServer from './api/comments.json';
-import usersFromServer from './api/users.json';
-
-function getUserById(userId) {
-  return usersFromServer.find(user => user.id === userId) || null;
-}
-
-function getCommentByPostId(postId) {
-  return commentsFromServer.filter(comment => comment.postId === postId);
-}
-
-const posts = postsFromServer.map(post => ({
-  ...post,
-  user: getUserById(post.userId),
-  comments: getCommentByPostId(post.id),
-}));
+import { PostList } from './components/PostList';
+import { getPosts } from './api';
 
 export const App = () => (
   <section className="App">
     <h1 className="App__title">Static list of posts</h1>
 
-    <PostList posts={posts} />
+    <PostList posts={getPosts()} />
   </section>
 );
